+++ src/plugins/explore/public/components/visualizations/line/line_chart_rules.ts
