+++ src/plugins/explore/public/components/visualizations/line/line_vis_options.tsx
