+++ src/plugins/explore/public/components/visualizations/style_panel/axes_options.tsx
