+++ src/plugins/explore/public/components/visualizations/style_panel/basic_vis_options.tsx
