+++ src/plugins/explore/public/components/visualizations/style_panel/grid_options.stories.tsx
