+++ src/plugins/explore/public/components/visualizations/style_panel/index.ts
