+++ src/plugins/explore/public/components/visualizations/style_panel/option.tsx
