+++ src/plugins/explore/public/components/visualizations/style_panel/threshold_options.tsx
