+++ src/plugins/explore/public/components/visualizations/utils/collections.ts
