+++ src/plugins/explore/public/components/visualizations/visualization.tsx
@@ -5,7 +5,7 @@
 
 import { EuiFlexGroup, EuiFlexItem, EuiPanel } from '@elastic/eui';
 import React from 'react';
-import { IExpressionLoaderParams } from '../../../../../../../../expressions/public';
+import { IExpressionLoaderParams } from '../../../../expressions/public';
 import { VisualizationTypeResult } from './utils/use_visualization_types';
 import { LineChartStyleControls } from './line/line_vis_config';
