+++ src/plugins/explore/public/components/visualizations/visualization_container.scss
