+++ src/plugins/explore/public/components/visualizations/visualization_registry.ts
